Refactor: 장소 검색 API 디렉토리 이동 (#69)
