+++ src/main/java/com/ilta/solepli/domain/sollect/service/SollectService.java
@@ -23,7 +23,6 @@ import com.ilta.solepli.domain.sollect.dto.PopularSollectResponseContent;
 import com.ilta.solepli.domain.sollect.dto.SollectSearchResponseContent;
 import com.ilta.solepli.domain.sollect.dto.request.SollectCreateRequest;
 import com.ilta.solepli.domain.sollect.dto.request.SollectUpdateRequest;
-import com.ilta.solepli.domain.sollect.dto.response.PlaceSearchResponse;
 import com.ilta.solepli.domain.sollect.dto.response.SollectCreateResponse;
 import com.ilta.solepli.domain.sollect.dto.response.SollectDetailResponse;
 import com.ilta.solepli.domain.sollect.dto.response.SollectPlaceAddPreviewResponse;
@@ -411,11 +410,6 @@ public class SollectService {
                 .hasNext(hasNext)
                 .build())
         .build();
-  }
-
-  @Transactional(readOnly = true)
-  public List<PlaceSearchResponse> getSearchPlaces(String keyword) {
-    return placeRepository.getPlacesByKeyword(keyword);
   }
 
   @Transactional(readOnly = true)
